requirements.txt
absl-py==1.4.0
antlr4-python3-runtime==4.9.3
black==23.3.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.34
-e git+https://github.com/facebookresearch/detectron2.git@e020497c85873c2b811ac87dd2e4a34a806e4c2b#egg=detectron2
einops==0.6.1
filelock==3.12.0
fonttools==4.39.3
fsspec==2023.4.0
fvcore==0.1.5.post20221221
google-auth==2.17.3
google-auth-oauthlib==1.0.0
grpcio==1.54.0
h5py==3.8.0
huggingface-hub==0.14.1
hydra-core==1.3.2
idna==3.4
imageio==2.28.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
iopath==0.1.9
kiwisolver==1.4.4
lazy_loader==0.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
MultiScaleDeformableAttention==1.0
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.72
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.0
portalocker==2.7.0
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.6
pygame==2.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0
requests==2.29.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.20.0
scipy==1.9.1
shapely==2.0.1
six==1.16.0
submitit==1.4.5
tabulate==0.9.0
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.3.0
tifffile==2023.4.12
timm==0.6.13
tomli==2.0.1
torch==1.9.0+cu111
torchaudio==0.9.0
torchvision==0.10.0+cu111
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.3.3
wget==3.2
yacs==0.1.8
zipp==3.15.0
